+++ source/extensions/lightspeed.trex.properties_pane.logic.widget/lightspeed/trex/properties_pane/logic/widget/__init__.py
@@ -17,4 +17,5 @@
 
 from .setup_ui import LogicPropertyWidget
 
+
 __all__ = ["LogicPropertyWidget"]