+++ source/extensions/lightspeed.trex.properties_pane.logic.widget/lightspeed/trex/properties_pane/logic/widget/tests/__init__.py
@@ -17,5 +17,6 @@
 
 from .e2e.test_prim_picker_ui import TestLogicWidgetUIComplete
 from .unit.test_prim_picker_integration import TestLogicWidgetIntegration
+from .unit.test_utils import TestGetOgnUiMetadata
 
-__all__ = ["TestLogicWidgetIntegration", "TestLogicWidgetUIComplete"]
+__all__ = ["TestGetOgnUiMetadata", "TestLogicWidgetIntegration", "TestLogicWidgetUIComplete"]