+++ source/extensions/omni.flux.property_widget_builder.delegates/omni/flux/property_widget_builder/delegates/int_value/__init__.py
@@ -15,6 +15,7 @@
 * limitations under the License.
 """
 
-__all__ = ("BytesToHuman",)
+__all__ = ("BytesToHuman", "IntSliderField")
 
 from .bytes_to_human_read import BytesToHuman
+from .slider import IntSliderField